+++ tests/layer_validate.rs
@@ -101,6 +101,24 @@ fn validate_doctl() {
 
 #[test]
 #[ignore]
+fn validate_ddog() {
+    validate("ddog");
+}
+
+#[test]
+#[ignore]
+fn validate_sntry() {
+    validate("sntry");
+}
+
+#[test]
+#[ignore]
+fn validate_secunit() {
+    validate("secunit");
+}
+
+#[test]
+#[ignore]
 fn validate_postgres() {
     validate("postgres");
 }